Add new image translate module then update image mode, and renamed some modules.
