+++ python_gui/UI/main_ui.py
@@ -39,8 +39,10 @@ class Ui_mainWindow(object):
         self.label_8.setObjectName("label_8")
         self.formLayout.setWidget(0, QtWidgets.QFormLayout.LabelRole, self.label_8)
         self.spinBox_baudrate = QtWidgets.QSpinBox(self.groupBox)
-        self.spinBox_baudrate.setMinimum(1000000)
-        self.spinBox_baudrate.setMaximum(1000000)
+        self.spinBox_baudrate.setMinimum(9600)
+        self.spinBox_baudrate.setMaximum(4000000)
+        self.spinBox_baudrate.setSingleStep(100)
+        self.spinBox_baudrate.setProperty("value", 1000000)
         self.spinBox_baudrate.setObjectName("spinBox_baudrate")
         self.formLayout.setWidget(2, QtWidgets.QFormLayout.FieldRole, self.spinBox_baudrate)
         self.label_2 = QtWidgets.QLabel(self.groupBox)
@@ -59,7 +61,7 @@ class Ui_mainWindow(object):
         self.line.setObjectName("line")
         self.verticalLayout.addWidget(self.line)
         self.groupBox_2 = QtWidgets.QGroupBox(self.widget)
-        self.groupBox_2.setEnabled(False)
+        self.groupBox_2.setEnabled(True)
         self.groupBox_2.setObjectName("groupBox_2")
         self.formLayout_2 = QtWidgets.QFormLayout(self.groupBox_2)
         self.formLayout_2.setObjectName("formLayout_2")
@@ -67,34 +69,60 @@ class Ui_mainWindow(object):
         self.label_3.setObjectName("label_3")
         self.formLayout_2.setWidget(0, QtWidgets.QFormLayout.LabelRole, self.label_3)
         self.spinBox_displayCol = QtWidgets.QSpinBox(self.groupBox_2)
-        self.spinBox_displayCol.setMinimum(128)
-        self.spinBox_displayCol.setMaximum(128)
+        self.spinBox_displayCol.setMinimum(8)
+        self.spinBox_displayCol.setMaximum(480)
+        self.spinBox_displayCol.setSingleStep(8)
+        self.spinBox_displayCol.setProperty("value", 128)
         self.spinBox_displayCol.setObjectName("spinBox_displayCol")
         self.formLayout_2.setWidget(0, QtWidgets.QFormLayout.FieldRole, self.spinBox_displayCol)
         self.label_4 = QtWidgets.QLabel(self.groupBox_2)
         self.label_4.setObjectName("label_4")
         self.formLayout_2.setWidget(1, QtWidgets.QFormLayout.LabelRole, self.label_4)
         self.spinBox_displayRow = QtWidgets.QSpinBox(self.groupBox_2)
-        self.spinBox_displayRow.setMinimum(64)
-        self.spinBox_displayRow.setMaximum(64)
+        self.spinBox_displayRow.setMinimum(8)
+        self.spinBox_displayRow.setMaximum(480)
+        self.spinBox_displayRow.setSingleStep(8)
+        self.spinBox_displayRow.setProperty("value", 64)
         self.spinBox_displayRow.setObjectName("spinBox_displayRow")
         self.formLayout_2.setWidget(1, QtWidgets.QFormLayout.FieldRole, self.spinBox_displayRow)
         self.label_5 = QtWidgets.QLabel(self.groupBox_2)
         self.label_5.setObjectName("label_5")
-        self.formLayout_2.setWidget(2, QtWidgets.QFormLayout.LabelRole, self.label_5)
-        self.comboBox_scanMode = QtWidgets.QComboBox(self.groupBox_2)
-        self.comboBox_scanMode.setObjectName("comboBox_scanMode")
-        self.comboBox_scanMode.addItem("")
-        self.comboBox_scanMode.addItem("")
-        self.formLayout_2.setWidget(2, QtWidgets.QFormLayout.FieldRole, self.comboBox_scanMode)
+        self.formLayout_2.setWidget(4, QtWidgets.QFormLayout.LabelRole, self.label_5)
+        self.comboBox_displayScanDirection = QtWidgets.QComboBox(self.groupBox_2)
+        self.comboBox_displayScanDirection.setObjectName("comboBox_displayScanDirection")
+        self.comboBox_displayScanDirection.addItem("")
+        self.comboBox_displayScanDirection.addItem("")
+        self.formLayout_2.setWidget(4, QtWidgets.QFormLayout.FieldRole, self.comboBox_displayScanDirection)
         self.label_6 = QtWidgets.QLabel(self.groupBox_2)
         self.label_6.setObjectName("label_6")
-        self.formLayout_2.setWidget(3, QtWidgets.QFormLayout.LabelRole, self.label_6)
-        self.comboBox_byteDirection = QtWidgets.QComboBox(self.groupBox_2)
-        self.comboBox_byteDirection.setObjectName("comboBox_byteDirection")
-        self.comboBox_byteDirection.addItem("")
-        self.comboBox_byteDirection.addItem("")
-        self.formLayout_2.setWidget(3, QtWidgets.QFormLayout.FieldRole, self.comboBox_byteDirection)
+        self.formLayout_2.setWidget(5, QtWidgets.QFormLayout.LabelRole, self.label_6)
+        self.comboBox_displayByteDirection = QtWidgets.QComboBox(self.groupBox_2)
+        self.comboBox_displayByteDirection.setObjectName("comboBox_displayByteDirection")
+        self.comboBox_displayByteDirection.addItem("")
+        self.comboBox_displayByteDirection.addItem("")
+        self.formLayout_2.setWidget(5, QtWidgets.QFormLayout.FieldRole, self.comboBox_displayByteDirection)
+        self.label_13 = QtWidgets.QLabel(self.groupBox_2)
+        self.label_13.setObjectName("label_13")
+        self.formLayout_2.setWidget(6, QtWidgets.QFormLayout.LabelRole, self.label_13)
+        self.comboBox_displaySignBit = QtWidgets.QComboBox(self.groupBox_2)
+        self.comboBox_displaySignBit.setObjectName("comboBox_displaySignBit")
+        self.comboBox_displaySignBit.addItem("")
+        self.comboBox_displaySignBit.addItem("")
+        self.formLayout_2.setWidget(6, QtWidgets.QFormLayout.FieldRole, self.comboBox_displaySignBit)
+        self.line_2 = QtWidgets.QFrame(self.groupBox_2)
+        self.line_2.setFrameShape(QtWidgets.QFrame.HLine)
+        self.line_2.setFrameShadow(QtWidgets.QFrame.Sunken)
+        self.line_2.setObjectName("line_2")
+        self.formLayout_2.setWidget(7, QtWidgets.QFormLayout.LabelRole, self.line_2)
+        self.label_14 = QtWidgets.QLabel(self.groupBox_2)
+        self.label_14.setObjectName("label_14")
+        self.formLayout_2.setWidget(2, QtWidgets.QFormLayout.LabelRole, self.label_14)
+        self.checkBox_displayFlipH = QtWidgets.QCheckBox(self.groupBox_2)
+        self.checkBox_displayFlipH.setObjectName("checkBox_displayFlipH")
+        self.formLayout_2.setWidget(2, QtWidgets.QFormLayout.FieldRole, self.checkBox_displayFlipH)
+        self.checkBox_displayFlipV = QtWidgets.QCheckBox(self.groupBox_2)
+        self.checkBox_displayFlipV.setObjectName("checkBox_displayFlipV")
+        self.formLayout_2.setWidget(3, QtWidgets.QFormLayout.FieldRole, self.checkBox_displayFlipV)
         self.verticalLayout.addWidget(self.groupBox_2)
         spacerItem = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
         self.verticalLayout.addItem(spacerItem)
@@ -167,7 +195,10 @@ class Ui_mainWindow(object):
         self.horizontalSlider_BWThreshold.setMinimum(1)
         self.horizontalSlider_BWThreshold.setMaximum(254)
         self.horizontalSlider_BWThreshold.setProperty("value", 127)
+        self.horizontalSlider_BWThreshold.setTracking(True)
         self.horizontalSlider_BWThreshold.setOrientation(QtCore.Qt.Horizontal)
+        self.horizontalSlider_BWThreshold.setInvertedAppearance(False)
+        self.horizontalSlider_BWThreshold.setInvertedControls(False)
         self.horizontalSlider_BWThreshold.setTickPosition(QtWidgets.QSlider.TicksBothSides)
         self.horizontalSlider_BWThreshold.setTickInterval(10)
         self.horizontalSlider_BWThreshold.setObjectName("horizontalSlider_BWThreshold")
@@ -561,12 +592,18 @@ class Ui_mainWindow(object):
         self.groupBox_2.setTitle(_translate("mainWindow", "Display Screen Setting"))
         self.label_3.setText(_translate("mainWindow", "Column"))
         self.label_4.setText(_translate("mainWindow", "Row"))
-        self.label_5.setText(_translate("mainWindow", "Scan Mode"))
-        self.comboBox_scanMode.setItemText(0, _translate("mainWindow", "Horizontal"))
-        self.comboBox_scanMode.setItemText(1, _translate("mainWindow", "Vertical"))
+        self.label_5.setText(_translate("mainWindow", "Scan Direction"))
+        self.comboBox_displayScanDirection.setItemText(0, _translate("mainWindow", "Horizontal"))
+        self.comboBox_displayScanDirection.setItemText(1, _translate("mainWindow", "Vertical"))
         self.label_6.setText(_translate("mainWindow", "Byte Direction"))
-        self.comboBox_byteDirection.setItemText(0, _translate("mainWindow", "Vertical"))
-        self.comboBox_byteDirection.setItemText(1, _translate("mainWindow", "Horizontal"))
+        self.comboBox_displayByteDirection.setItemText(0, _translate("mainWindow", "Vertical"))
+        self.comboBox_displayByteDirection.setItemText(1, _translate("mainWindow", "Horizontal"))
+        self.label_13.setText(_translate("mainWindow", "Significan Bit"))
+        self.comboBox_displaySignBit.setItemText(0, _translate("mainWindow", "MSB"))
+        self.comboBox_displaySignBit.setItemText(1, _translate("mainWindow", "LSB"))
+        self.label_14.setText(_translate("mainWindow", "Flip Mode"))
+        self.checkBox_displayFlipH.setText(_translate("mainWindow", "Horizontal"))
+        self.checkBox_displayFlipV.setText(_translate("mainWindow", "Vertical"))
         self.textEdit.setHtml(_translate("mainWindow", "<!DOCTYPE HTML PUBLIC \"-//W3C//DTD HTML 4.0//EN\" \"http://www.w3.org/TR/REC-html40/strict.dtd\">\n"
 "<html><head><meta name=\"qrichtext\" content=\"1\" /><style type=\"text/css\">\n"
 "p, li { white-space: pre-wrap; }\n"
