+++ python_gui/UI/main_ui.ui
@@ -56,9 +56,15 @@
         <item row="2" column="1">
          <widget class="QSpinBox" name="spinBox_baudrate">
           <property name="minimum">
-           <number>1000000</number>
+           <number>9600</number>
           </property>
           <property name="maximum">
+           <number>4000000</number>
+          </property>
+          <property name="singleStep">
+           <number>100</number>
+          </property>
+          <property name="value">
            <number>1000000</number>
           </property>
          </widget>
@@ -93,7 +99,7 @@
      <item>
       <widget class="QGroupBox" name="groupBox_2">
        <property name="enabled">
-        <bool>false</bool>
+        <bool>true</bool>
        </property>
        <property name="title">
         <string>Display Screen Setting</string>
@@ -109,9 +115,15 @@
         <item row="0" column="1">
          <widget class="QSpinBox" name="spinBox_displayCol">
           <property name="minimum">
-           <number>128</number>
+           <number>8</number>
           </property>
           <property name="maximum">
+           <number>480</number>
+          </property>
+          <property name="singleStep">
+           <number>8</number>
+          </property>
+          <property name="value">
            <number>128</number>
           </property>
          </widget>
@@ -126,22 +138,28 @@
         <item row="1" column="1">
          <widget class="QSpinBox" name="spinBox_displayRow">
           <property name="minimum">
-           <number>64</number>
+           <number>8</number>
           </property>
           <property name="maximum">
+           <number>480</number>
+          </property>
+          <property name="singleStep">
+           <number>8</number>
+          </property>
+          <property name="value">
            <number>64</number>
           </property>
          </widget>
         </item>
-        <item row="2" column="0">
+        <item row="4" column="0">
          <widget class="QLabel" name="label_5">
           <property name="text">
-           <string>Scan Mode</string>
+           <string>Scan Direction</string>
           </property>
          </widget>
         </item>
-        <item row="2" column="1">
-         <widget class="QComboBox" name="comboBox_scanMode">
+        <item row="4" column="1">
+         <widget class="QComboBox" name="comboBox_displayScanDirection">
           <item>
            <property name="text">
             <string>Horizontal</string>
@@ -154,15 +172,15 @@
           </item>
          </widget>
         </item>
-        <item row="3" column="0">
+        <item row="5" column="0">
          <widget class="QLabel" name="label_6">
           <property name="text">
            <string>Byte Direction</string>
           </property>
          </widget>
         </item>
-        <item row="3" column="1">
-         <widget class="QComboBox" name="comboBox_byteDirection">
+        <item row="5" column="1">
+         <widget class="QComboBox" name="comboBox_displayByteDirection">
           <item>
            <property name="text">
             <string>Vertical</string>
@@ -173,6 +191,55 @@
             <string>Horizontal</string>
            </property>
           </item>
+         </widget>
+        </item>
+        <item row="6" column="0">
+         <widget class="QLabel" name="label_13">
+          <property name="text">
+           <string>Significan Bit</string>
+          </property>
+         </widget>
+        </item>
+        <item row="6" column="1">
+         <widget class="QComboBox" name="comboBox_displaySignBit">
+          <item>
+           <property name="text">
+            <string>MSB</string>
+           </property>
+          </item>
+          <item>
+           <property name="text">
+            <string>LSB</string>
+           </property>
+          </item>
+         </widget>
+        </item>
+        <item row="7" column="0">
+         <widget class="Line" name="line_2">
+          <property name="orientation">
+           <enum>Qt::Horizontal</enum>
+          </property>
+         </widget>
+        </item>
+        <item row="2" column="0">
+         <widget class="QLabel" name="label_14">
+          <property name="text">
+           <string>Flip Mode</string>
+          </property>
+         </widget>
+        </item>
+        <item row="2" column="1">
+         <widget class="QCheckBox" name="checkBox_displayFlipH">
+          <property name="text">
+           <string>Horizontal</string>
+          </property>
+         </widget>
+        </item>
+        <item row="3" column="1">
+         <widget class="QCheckBox" name="checkBox_displayFlipV">
+          <property name="text">
+           <string>Vertical</string>
+          </property>
          </widget>
         </item>
        </layout>
@@ -335,8 +402,17 @@ p, li { white-space: pre-wrap; }
                <property name="value">
                 <number>127</number>
                </property>
+               <property name="tracking">
+                <bool>true</bool>
+               </property>
                <property name="orientation">
                 <enum>Qt::Horizontal</enum>
+               </property>
+               <property name="invertedAppearance">
+                <bool>false</bool>
+               </property>
+               <property name="invertedControls">
+                <bool>false</bool>
                </property>
                <property name="tickPosition">
                 <enum>QSlider::TicksBothSides</enum>
